feat: support 64-bit addressing for virtio-blk
- Support 64-bit physical addresses for VirtQueue descriptors, avail, and used rings.
- Update MMIO logic to handle 32-bit Low/High register pairs using bitmasks to
  ensure partial updates don't overwrite the full 64-bit address.

+++ include/device/virtio_blk.hpp
@@ -131,9 +131,9 @@ struct VblkReqHeader {
 
 struct VirtioBlkQueue {
     uint32_t queue_num = 0;
-    uint32_t queue_desc = 0;
-    uint32_t queue_avail = 0;
-    uint32_t queue_used = 0;
+    uint64_t queue_desc = 0;
+    uint64_t queue_avail = 0;
+    uint64_t queue_used = 0;
     uint16_t last_avail = 0;
     bool ready = false;
 };
